implement confirmation and reset working, only redirect route has to be given by shell

+++ auth/confirmation.js
@@ -3,6 +3,7 @@
 const {sendConfirmationEmail,sendPasswordResetEmail} = require ('./email');
 const UserTokenModel = require ('../models/userToken');
 const UserModel = require ('../models/user');
+const {randomToken} = require('../encryption/encryption');
 
 const startConfirmationProcess = async (user) => {
     const userId = user._id;
@@ -21,9 +22,15 @@ const startConfirmationProcess = async (user) => {
     await sendConfirmationEmail(userEmail,userId,confirmationToken);
 }
 
-const startResetProcess = async (user) => {
+const startResetProcess = async (email) => {
+    const user = await UserModel.findOne({email: email});
+
+    if(user === null) {
+        return false;
+    }
+
     const userId = user._id;
-    const userEmail = user.email;
+    const userName = user.firstName + " " + user.lastName;
 
     // delete old tokens, create new one
     await UserTokenModel.deleteMany({userID: userId, tokenType: "PASSWORD_RESET" });
@@ -34,7 +41,8 @@ const startResetProcess = async (user) => {
     const confirmationToken = await userToken.token;
 
     // send email with token
-    // await sendPasswordResetEmail(userEmail,userId,confirmationToken);
+    await sendPasswordResetEmail(email,userId,userName,confirmationToken);
+    return true;
 }
 
 const endConfirmationProcess = async (id, token) => {
@@ -63,12 +71,41 @@ const endConfirmationProcess = async (id, token) => {
     }
 }
 
-const endResetProcess = async (user, token) => {
+// middle call between forgot and change -> confirm
+const resetConfirm = async (id, token) => {
 
+    const userToken = await UserTokenModel.findOne({userID: id, tokenType: "PASSWORD_RESET"})
+    if(userToken === null)
+        return false;
+    else if(userToken.token !== token)
+        return false;
+    else if(userToken.expirationDate < new Date().getTime())
+        return false;
+    else {
+        const newToken = await randomToken();
+        await UserTokenModel.updateOne({userID: id, tokenType: "PASSWORD_RESET"}, {token: newToken}, null);
+        return {token: newToken};
+    }
+}
+
+const endResetProcess = async (id, token) => {
+
+    const databaseToken = await UserTokenModel.findOne({userID:id, token: token, tokenType: "PASSWORD_RESET"})
+
+    if(databaseToken === null)
+        return false;
+    else if (databaseToken.expirationDate < new Date().getTime())
+        return false;
+    else {
+        await UserTokenModel.deleteMany({userID:id, tokenType: "PASSWORD_RESET"});
+        return true;
+    }
 }
 
 module.exports = {
     startConfirmationProcess,
     startResetProcess,
-    endConfirmationProcess
+    endConfirmationProcess,
+    resetConfirm,
+    endResetProcess
 }
